added all props andstate and passed all tests
