+++ src/Components/Attendees.js
@@ -1,3 +1,23 @@
-export default function Attendees() {
-  return;
+
+import Attendee from "./Attendee"
+
+export default function Attendees({updateEventAttendance,event, attendees}) {
+  // setShowAttendees(false)
+  // const {people: attendees} = event
+
+
+  return(
+  <div className="attendees">
+    {attendees.map((attendee, index) => (
+    <>
+      <Attendee 
+  
+      event={event} attendee={attendee}
+      updateEventAttendance={updateEventAttendance}/>
+     </>
+  
+    ))}
+  </div>
+
+  )
 }
