+++ src/Components/Event.js
@@ -1,3 +1,42 @@
-export default function Event() {
-  return;
+import { useState } from "react";
+import Attendees from "./Attendees";
+
+
+export default function Event({event, attendees, updateEventAttendance}) {
+
+  const [showAttendees, setShowAttendees] = useState(false);
+  // const attendees = event.people.id
+
+  function toggleEventAttendees() {
+    setShowAttendees(!showAttendees);
+  }
+
+  // const index = attendees.people.indexOf(event.id)
+  return(
+<div className="event">
+  <>
+    <li key={event.id}>
+    <img src={event.eventImage} alt={event.name} />
+    <h5>
+      {event.name} {event.eventType}
+    </h5>
+    <br />
+    <span>Organized by: {event.organizer} </span>
+    <br />
+
+     <>
+      <button onClick={toggleEventAttendees}>
+        {!showAttendees ? "Show Attendees" : "Hide Attendees"}
+      </button>
+
+      {showAttendees ? ( <Attendees
+
+      attendees = {attendees}
+      event = {event}
+      updateEventAttendance={updateEventAttendance}/> ) : null}
+    </>
+  </li>
+  </>
+  </div>
+  )
 }
